backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
python-jose[cryptography]==3.3.0
google-generativeai==0.8.5
langchain-together==0.2.0
langchain-groq==0.2.3
pyyaml==6.0.2
fpdf2==2.8.2
pytest==8.3.4
pytest-asyncio==0.25.1
